Add files via upload
It is just my practical case , the effectiveness of a web scraper heavily depends on the specific URL and the underlying structure of the webpage. Issues such as raw HTML or embedded JavaScript can arise, complicating data extraction. Consequently, to create an effective scraping solution for a particular site, the program must be tailored to that specific page rather than attempting to generalize for multiple sites. This targeted approach ensures that the scraper can navigate the unique structure and dynamic elements of the website, maximizing data accuracy and relevance.